train and test VAD pipeline
