+++ fsd/agents/VAD/VAD_head.py
@@ -1,8 +1,10 @@
 import copy
-
+import numpy as np
 import torch
 import torch.nn as nn
+from torch.nn.init import normal_
 import torch.nn.functional as F
+from torchvision.transforms.functional import rotate
 
 from mmengine.utils import digit_version
 TORCH_VERSION = tuple(int(x) for x in torch.__version__.split('.')[:2])
@@ -18,6 +20,8 @@ from mmdet3d.models.task_modules.builder import build_bbox_coder
 
 from fsd.registry import TASK_UTILS, MODELS, HEADS
 from fsd.models.detectors.bevformer.task_modules.bbox.util import normalize_bbox 
+from fsd.models import MultiScaleDeformableAttention3D, TemporalSelfAttention
+from fsd.models.detectors.bevformer.modules.decoder import BEVMultiScaleDeformableAttention
 
 from .utils import get_traj_warmup_loss_weight
 from .utils import (normalize_2d_pts, 
@@ -92,7 +96,18 @@ class VADHead(DETRHead):
                      normalize=True),
                  with_box_refine=False,
                  as_two_stage=False,
-                 transformer=None,
+                 #transformer=None,
+                 num_feature_levels=4,
+                 num_cams=6,
+                 bev_encoder = None,
+                 bev_decoder = None,
+                 map_decoder = None,
+                 rotate_prev_bev=True,
+                 use_shift=True,
+                 use_can_bus=True,
+                 can_bus_norm=True,
+                 use_cams_embeds=True,
+                 rotate_center=[100, 100],
                  bbox_coder=None,
                  num_cls_fcs=2,
                  code_weights=None,
@@ -153,11 +168,19 @@ class VADHead(DETRHead):
                  ego_lcf_feat_idx=None,
                  valid_fut_ts=6,
                  **kwargs):
-        #TODO: what is this query?
+    
         self.num_query = num_query
         self.bev_h = bev_h
         self.bev_w = bev_w
-        self.fp16_enabled = False
+        self.num_feature_levels = num_feature_levels
+        self.num_cams = num_cams
+        self.rotate_prev_bev = rotate_prev_bev
+        self.use_shift = use_shift
+        self.use_can_bus = use_can_bus
+        self.can_bus_norm = can_bus_norm
+        self.use_cams_embeds = use_cams_embeds
+        self.rotate_center = rotate_center
+        
         self.fut_ts = fut_ts
         self.fut_mode = fut_mode
         self.tot_epoch = tot_epoch
@@ -171,6 +194,9 @@ class VADHead(DETRHead):
         self.pe_normalization = pe_normalization
         self.ego_his_encoder = ego_his_encoder
         self.ego_fut_mode = ego_fut_mode
+        self.bev_encoder = bev_encoder
+        self.bev_decoder = bev_decoder
+        self.map_decoder = map_decoder
         self.ego_agent_decoder = ego_agent_decoder
         self.ego_map_decoder = ego_map_decoder
         self.query_thresh = query_thresh
@@ -185,8 +211,8 @@ class VADHead(DETRHead):
 
         self.with_box_refine = with_box_refine
         self.as_two_stage = as_two_stage
-        if self.as_two_stage:
-            transformer['as_two_stage'] = self.as_two_stage
+        #if self.as_two_stage:
+        #    transformer['as_two_stage'] = self.as_two_stage
         if 'code_size' in kwargs:
             self.code_size = kwargs['code_size']
         else:
@@ -278,11 +304,15 @@ class VADHead(DETRHead):
         self.num_decoder_layers = 1
         self.num_map_decoder_layers = 1
         self.num_motion_decoder_layers = 1
-        if transformer.decoder is not None:
-            self.num_decoder_layers = transformer.decoder.num_layers
-        if transformer.map_decoder is not None:
-            self.num_map_decoder_layers = transformer.map_decoder.num_layers
-            
+        #if transformer.decoder is not None:
+        #    self.num_decoder_layers = transformer.decoder.num_layers
+        #if transformer.map_decoder is not None:
+        #    self.num_map_decoder_layers = transformer.map_decoder.num_layers
+        if self.bev_decoder is not None:
+            self.num_decoder_layers = self.bev_decoder.num_layers
+        if self.map_decoder is not None:
+            self.num_map_decoder_layers = self.map_decoder.num_layers
+        
         # init()
         super(VADHead, self).__init__(*args, **kwargs)
         self.code_weights = nn.Parameter(torch.tensor(
@@ -291,7 +321,15 @@ class VADHead(DETRHead):
             self.map_code_weights, requires_grad=False), requires_grad=False)
         
         # build transformer
-        self.transformer = MODELS.build(transformer)
+        #self.transformer = MODELS.build(transformer)
+        self.bev_encoder = MODELS.build(bev_encoder)
+        self.bev_decoder = MODELS.build(bev_decoder)
+        self.map_decoder = MODELS.build(map_decoder)
+        self.motion_decoder = MODELS.build(motion_decoder)
+        self.motion_map_decoder = MODELS.build(motion_map_decoder)
+        self.ego_agent_decoder = MODELS.build(ego_agent_decoder)
+        self.ego_map_decoder = MODELS.build(ego_map_decoder)
+        
         # build positional encoding
         self.positional_encoding = MODELS.build(positional_encoding) # bev positional encoding
         
@@ -333,6 +371,7 @@ class VADHead(DETRHead):
         traj_branch.append(Linear(self.embed_dims*2, self.fut_ts*2))
         traj_branch = nn.Sequential(*traj_branch)
 
+        # turning left, straight, turning right or k modes
         traj_cls_branch = []
         for _ in range(self.num_reg_fcs):
             traj_cls_branch.append(Linear(self.embed_dims*2, self.embed_dims*2))
@@ -403,8 +442,25 @@ class VADHead(DETRHead):
                 self.map_instance_embedding = nn.Embedding(self.map_num_vec, self.embed_dims * 2)
                 self.map_pts_embedding = nn.Embedding(self.map_num_pts_per_vec, self.embed_dims * 2)
         
+        # bevformer
+        self.level_embeds = nn.Parameter(torch.Tensor(
+            self.num_feature_levels, self.embed_dims))
+        self.cams_embeds = nn.Parameter(
+            torch.Tensor(self.num_cams, self.embed_dims))
+        self.reference_points = nn.Linear(self.embed_dims, 3)
+        self.map_reference_points = nn.Linear(self.embed_dims, 2)
+        self.can_bus_mlp = nn.Sequential(
+            nn.Linear(18, self.embed_dims // 2),
+            nn.ReLU(inplace=True),
+            nn.Linear(self.embed_dims // 2, self.embed_dims),
+            nn.ReLU(inplace=True),
+        )
+        if self.can_bus_norm:
+            self.can_bus_mlp.add_module('norm', nn.LayerNorm(self.embed_dims))
+        
+        # agent/agent, agent/map interaction
         if self.motion_decoder is not None:
-            self.motion_decoder = build_transformer_layer_sequence(self.motion_decoder)
+            #self.motion_decoder = build_transformer_layer_sequence(self.motion_decoder)
             self.motion_mode_query = nn.Embedding(self.fut_mode, self.embed_dims)	
             self.motion_mode_query.weight.requires_grad = True
             if self.use_pe:
@@ -417,7 +473,7 @@ class VADHead(DETRHead):
             self.motion_map_decoder = build_transformer_layer_sequence(self.motion_map_decoder)
             if self.use_pe:
                 self.pos_mlp = nn.Linear(2, self.embed_dims)
-        
+        # ego
         if self.ego_his_encoder is not None:
             self.ego_his_encoder = LaneNet(2, self.embed_dims//2, 3)
         else:
@@ -448,9 +504,19 @@ class VADHead(DETRHead):
             nn.ReLU(),
             nn.Linear(self.embed_dims, self.embed_dims, bias=True))
 
+    def _init_attention_modules(self, modules):
+        """Initialize attention modules."""
+        for m in modules:
+            if isinstance(m, MultiScaleDeformableAttention3D) or isinstance(m, TemporalSelfAttention) \
+                    or isinstance(m, BEVMultiScaleDeformableAttention):
+                try:
+                    m.init_weight()
+                except AttributeError:
+                    m.init_weights()
+        
     def init_weights(self):
         """Initialize weights of the DeformDETR head."""
-        self.transformer.init_weights()
+        #self.transformer.init_weights()
         if self.loss_cls.use_sigmoid:
             bias_init = bias_init_with_prob(0.01)
             for m in self.cls_branches:
@@ -466,6 +532,29 @@ class VADHead(DETRHead):
         # for m in self.map_reg_branches:
         #     constant_init(m[-1], 0, bias=0)
         # nn.init.constant_(self.map_reg_branches[0][-1].bias.data[2:], 0.)
+        if self.bev_encoder is not None:
+            for p in self.bev_encoder.parameters():
+                if p.dim() > 1:
+                    nn.init.xavier_uniform_(p)
+            self._init_attention_modules(self.bev_encoder.modules())
+        
+        if self.bev_decoder is not None:
+            for p in self.bev_decoder.parameters():
+                if p.dim() > 1:
+                    nn.init.xavier_uniform_(p)
+            self._init_attention_modules(self.bev_decoder.modules())
+        
+        if self.map_decoder is not None:
+            for p in self.map_decoder.parameters():
+                if p.dim() > 1:
+                    nn.init.xavier_uniform_(p)
+            self._init_attention_modules(self.map_decoder.modules())
+        normal_(self.level_embeds)
+        normal_(self.cams_embeds)
+        xavier_init(self.reference_points, distribution='uniform', bias=0.)
+        xavier_init(self.map_reference_points, distribution='uniform', bias=0.)
+        xavier_init(self.can_bus_mlp, distribution='uniform', bias=0.)
+        
         if self.motion_decoder is not None:
             for p in self.motion_decoder.parameters():
                 if p.dim() > 1:
@@ -494,6 +583,220 @@ class VADHead(DETRHead):
             for p in self.ego_map_decoder.parameters():
                 if p.dim() > 1:
                     nn.init.xavier_uniform_(p)
+
+    def get_bev_features(
+            self,
+            mlvl_feats,
+            bev_queries,
+            bev_h,
+            bev_w,
+            grid_length=[0.512, 0.512],
+            bev_pos=None,
+            prev_bev=None,
+            **kwargs):
+        """
+        obtain bev features.
+        """
+
+        bs = mlvl_feats[0].size(0)
+        bev_queries = bev_queries.unsqueeze(1).repeat(1, bs, 1)
+        bev_pos = bev_pos.flatten(2).permute(2, 0, 1)
+
+        # obtain rotation angle and shift with ego motion
+        delta_x = np.array([each['bev_attr'][0]
+                           for each in kwargs['img_metas']])
+        delta_y = np.array([each['bev_attr'][1]
+                           for each in kwargs['img_metas']])
+        ego_angle = np.array(
+            [each['bev_attr'][-2] / np.pi * 180 for each in kwargs['img_metas']])
+        grid_length_y = grid_length[0]
+        grid_length_x = grid_length[1]
+        translation_length = np.sqrt(delta_x ** 2 + delta_y ** 2)
+        translation_angle = np.arctan2(delta_y, delta_x) / np.pi * 180
+        bev_angle = ego_angle - translation_angle
+        shift_y = translation_length * \
+            np.cos(bev_angle / 180 * np.pi) / grid_length_y / bev_h
+        shift_x = translation_length * \
+            np.sin(bev_angle / 180 * np.pi) / grid_length_x / bev_w
+        shift_y = shift_y * self.use_shift
+        shift_x = shift_x * self.use_shift
+        shift = bev_queries.new_tensor(
+            [shift_x, shift_y]).permute(1, 0)  # xy, bs -> bs, xy
+
+        if prev_bev is not None:
+            if prev_bev.shape[1] == bev_h * bev_w:
+                prev_bev = prev_bev.permute(1, 0, 2)
+            if self.rotate_prev_bev:
+                for i in range(bs):
+                    # num_prev_bev = prev_bev.size(1)
+                    rotation_angle = kwargs['img_metas'][i]['bev_attr'][-1]
+                    tmp_prev_bev = prev_bev[:, i].reshape(
+                        bev_h, bev_w, -1).permute(2, 0, 1)
+                    tmp_prev_bev = rotate(tmp_prev_bev, float(rotation_angle),
+                                          center=self.rotate_center)
+                    tmp_prev_bev = tmp_prev_bev.permute(1, 2, 0).reshape(
+                        bev_h * bev_w, 1, -1)
+                    prev_bev[:, i] = tmp_prev_bev[:, 0]
+
+        # add can bus signals
+        can_bus = bev_queries.new_tensor(
+            [each['bev_attr'] for each in kwargs['img_metas']])  # [:, :]
+        can_bus = self.can_bus_mlp(can_bus)[None, :, :]
+        bev_queries = bev_queries + can_bus * self.use_can_bus
+
+        feat_flatten = []
+        spatial_shapes = []
+        for lvl, feat in enumerate(mlvl_feats):
+            bs, num_cam, c, h, w = feat.shape
+            spatial_shape = (h, w)
+            feat = feat.flatten(3).permute(1, 0, 3, 2)
+            if self.use_cams_embeds:
+                feat = feat + self.cams_embeds[:, None, None, :].to(feat.dtype)
+            feat = feat + self.level_embeds[None,
+                                            None, lvl:lvl + 1, :].to(feat.dtype)
+            spatial_shapes.append(spatial_shape)
+            feat_flatten.append(feat)
+
+        feat_flatten = torch.cat(feat_flatten, 2)
+        spatial_shapes = torch.as_tensor(
+            spatial_shapes, dtype=torch.long, device=bev_pos.device)
+        level_start_index = torch.cat((spatial_shapes.new_zeros(
+            (1,)), spatial_shapes.prod(1).cumsum(0)[:-1]))
+
+        feat_flatten = feat_flatten.permute(
+            0, 2, 1, 3)  # (num_cam, H*W, bs, embed_dims)
+
+        bev_embed = self.bev_encoder(
+            bev_queries,
+            feat_flatten,
+            feat_flatten,
+            bev_h=bev_h,
+            bev_w=bev_w,
+            bev_pos=bev_pos,
+            spatial_shapes=spatial_shapes,
+            level_start_index=level_start_index,
+            prev_bev=prev_bev,
+            shift=shift,
+            **kwargs
+        )
+
+        return bev_embed
+
+    def get_bev_map_latent(self,
+                mlvl_feats,
+                bev_embed,
+                bev_queries,
+                object_query_embed,
+                map_query_embed,
+                bev_h,
+                bev_w,
+                grid_length=[0.512, 0.512],
+                bev_pos=None,
+                reg_branches=None,
+                cls_branches=None,
+                map_reg_branches=None,
+                map_cls_branches=None,                
+                prev_bev=None,            
+                **kwargs):
+        """Forward function for `Detr3DTransformer`.
+        Args:
+            mlvl_feats (list(Tensor)): Input queries from
+                different level. Each element has shape
+                [bs, num_cams, embed_dims, h, w].
+            bev_queries (Tensor): (bev_h*bev_w, c)
+            bev_pos (Tensor): (bs, embed_dims, bev_h, bev_w)
+            object_query_embed (Tensor): The query embedding for decoder,
+                with shape [num_query, c].
+            reg_branches (obj:`nn.ModuleList`): Regression heads for
+                feature maps from each decoder layer. Only would
+                be passed when `with_box_refine` is True. Default to None.
+        Returns:
+            tuple[Tensor]: results of decoder containing the following tensor.
+                - bev_embed: BEV features
+                - inter_states: Outputs from decoder. If
+                    return_intermediate_dec is True output has shape \
+                      (num_dec_layers, bs, num_query, embed_dims), else has \
+                      shape (1, bs, num_query, embed_dims).
+                - init_reference_out: The initial value of reference \
+                    points, has shape (bs, num_queries, 4).
+                - inter_references_out: The internal value of reference \
+                    points in decoder, has shape \
+                    (num_dec_layers, bs,num_query, embed_dims)
+                - enc_outputs_class: The classification score of \
+                    proposals generated from \
+                    encoder's feature maps, has shape \
+                    (batch, h*w, num_classes). \
+                    Only would be returned when `as_two_stage` is True, \
+                    otherwise None.
+                - enc_outputs_coord_unact: The regression results \
+                    generated from encoder's feature maps., has shape \
+                    (batch, h*w, 4). Only would \
+                    be returned when `as_two_stage` is True, \
+                    otherwise None.
+        """
+
+        bs = mlvl_feats[0].size(0)
+        query_pos, query = torch.split(
+            object_query_embed, self.embed_dims, dim=1)
+        query_pos = query_pos.unsqueeze(0).expand(bs, -1, -1)
+        query = query.unsqueeze(0).expand(bs, -1, -1)
+        reference_points = self.reference_points(query_pos)
+        reference_points = reference_points.sigmoid()
+        init_reference_out = reference_points
+
+        map_query_pos, map_query = torch.split(
+            map_query_embed, self.embed_dims, dim=1)
+        map_query_pos = map_query_pos.unsqueeze(0).expand(bs, -1, -1)
+        map_query = map_query.unsqueeze(0).expand(bs, -1, -1)
+        map_reference_points = self.map_reference_points(map_query_pos)
+        map_reference_points = map_reference_points.sigmoid()
+        map_init_reference_out = map_reference_points        
+
+        query = query.permute(1, 0, 2)
+        query_pos = query_pos.permute(1, 0, 2)
+        map_query = map_query.permute(1, 0, 2)
+        map_query_pos = map_query_pos.permute(1, 0, 2)
+        bev_embed = bev_embed.permute(1, 0, 2)
+
+        if self.bev_decoder is not None:
+            # [L, Q, B, D], [L, B, Q, D]
+            inter_states, inter_references = self.bev_decoder(
+                query=query,
+                key=None,
+                value=bev_embed,
+                query_pos=query_pos,
+                reference_points=reference_points,
+                reg_branches=reg_branches,
+                cls_branches=cls_branches,
+                spatial_shapes=torch.tensor([[bev_h, bev_w]], device=query.device),
+                level_start_index=torch.tensor([0], device=query.device),
+                **kwargs)
+            inter_references_out = inter_references
+        else:
+            inter_states = query.unsqueeze(0)
+            inter_references_out = reference_points.unsqueeze(0)
+
+        if self.map_decoder is not None:
+            # [L, Q, B, D], [L, B, Q, D]
+            map_inter_states, map_inter_references = self.map_decoder(
+                query=map_query,
+                key=None,
+                value=bev_embed,
+                query_pos=map_query_pos,
+                reference_points=map_reference_points,
+                reg_branches=map_reg_branches,
+                cls_branches=map_cls_branches,
+                spatial_shapes=torch.tensor([[bev_h, bev_w]], device=map_query.device),
+                level_start_index=torch.tensor([0], device=map_query.device),
+                **kwargs)
+            map_inter_references_out = map_inter_references
+        else:
+            map_inter_states = map_query.unsqueeze(0)
+            map_inter_references_out = map_reference_points.unsqueeze(0)
+
+        return (
+            bev_embed, inter_states, init_reference_out, inter_references_out,
+            map_inter_states, map_init_reference_out, map_inter_references_out)
 
     def forward(self,
                 mlvl_feats,
@@ -542,38 +845,40 @@ class VADHead(DETRHead):
         bev_mask = torch.zeros((bs, self.bev_h, self.bev_w),
                                device=bev_queries.device).to(dtype)
         bev_pos = self.positional_encoding(bev_mask).to(dtype)
-            
-        if only_bev:  # only use encoder to obtain BEV features, TODO: refine the workaround
-            return self.transformer.get_bev_features(
-                mlvl_feats,
-                bev_queries,
-                self.bev_h,
-                self.bev_w,
-                grid_length=(self.real_h / self.bev_h,
-                             self.real_w / self.bev_w),
-                bev_pos=bev_pos,
-                img_metas=img_metas,
-                prev_bev=prev_bev,
-            )
-        else:
-            #NOTE: this transformer is an instance of VADPerceptionTransformer, which includes BEVFormer as encoder, 
-            # a detection transformer as decoder, and a MapDetectionTransformer as map decoder.
-            outputs = self.transformer(
-                mlvl_feats,
-                bev_queries,
-                object_query_embeds,
-                map_query_embeds,
-                self.bev_h,
-                self.bev_w,
-                grid_length=(self.real_h / self.bev_h,
-                             self.real_w / self.bev_w),
-                bev_pos=bev_pos,
-                reg_branches=self.reg_branches if self.with_box_refine else None,  # noqa:E501
-                cls_branches=self.cls_branches if self.as_two_stage else None,
-                map_reg_branches=self.map_reg_branches if self.with_box_refine else None,  # noqa:E501
-                map_cls_branches=self.map_cls_branches if self.as_two_stage else None,
-                img_metas=img_metas,
-                prev_bev=prev_bev
+        
+        # if only return bev features, return it directly
+        bev_embed = self.get_bev_features(
+            mlvl_feats,
+            bev_queries,
+            self.bev_h,
+            self.bev_w,
+            grid_length=(self.real_h / self.bev_h,
+                            self.real_w / self.bev_w),
+            bev_pos=bev_pos,
+            img_metas=img_metas,
+            prev_bev=prev_bev,
+        )
+
+        if only_bev:
+            return bev_embed
+        
+        outputs = self.get_bev_map_latent(
+            mlvl_feats,
+            bev_embed,
+            bev_queries,
+            object_query_embeds,
+            map_query_embeds,
+            self.bev_h,
+            self.bev_w,
+            grid_length=(self.real_h / self.bev_h,
+                            self.real_w / self.bev_w),
+            bev_pos=bev_pos,
+            reg_branches=self.reg_branches if self.with_box_refine else None,  # noqa:E501
+            cls_branches=self.cls_branches if self.as_two_stage else None,
+            map_reg_branches=self.map_reg_branches if self.with_box_refine else None,  # noqa:E501
+            map_cls_branches=self.map_cls_branches if self.as_two_stage else None,
+            img_metas=img_metas,
+            prev_bev=prev_bev
         )
 
         bev_embed, hs, init_reference, inter_references, \
