+++ fsd/agents/VAD/VAD_transformer.py
@@ -30,8 +30,8 @@ def inverse_sigmoid(x, eps=1e-5):
 
 
 @MODELS.register_module()
-class MapDetectionTransformerDecoder(TransformerLayerSequence):
-    """Implements the decoder in DETR3D transformer.
+class VectorMapDecoder(TransformerLayerSequence):
+    """Implements the decoder in MapTR/VAD
     Args:
         return_intermediate (bool): Whether to return intermediate outputs.
         coder_norm_cfg (dict): Config of last normalization layer. Default:
@@ -39,9 +39,8 @@ class MapDetectionTransformerDecoder(TransformerLayerSequence):
     """
 
     def __init__(self, *args, return_intermediate=False, **kwargs):
-        super(MapDetectionTransformerDecoder, self).__init__(*args, **kwargs)
+        super(VectorMapDecoder, self).__init__(*args, **kwargs)
         self.return_intermediate = return_intermediate
-        self.fp16_enabled = False
 
     def forward(self,
                 query,
@@ -108,335 +107,17 @@ class MapDetectionTransformerDecoder(TransformerLayerSequence):
 
         return output, reference_points
 
-
 @MODELS.register_module()
-class VADPerceptionTransformer(BaseModule):
-    """Implements the Detr3D transformer.
-    Args:
-        as_two_stage (bool): Generate query from encoder features.
-            Default: False.
-        num_feature_levels (int): Number of feature maps from FPN:
-            Default: 4.
-        two_stage_num_proposals (int): Number of proposals when set
-            `as_two_stage` as True. Default: 300.
-    """
-
-    def __init__(self,
-                 num_feature_levels=4,
-                 num_cams=6,
-                 two_stage_num_proposals=300,
-                 encoder=None,
-                 decoder=None,
-                 map_decoder=None,
-                 embed_dims=256,
-                 rotate_prev_bev=True,
-                 use_shift=True,
-                 use_can_bus=True,
-                 can_bus_norm=True,
-                 use_cams_embeds=True,
-                 rotate_center=[100, 100],
-                 map_num_vec=50,
-                 map_num_pts_per_vec=10,
-                 **kwargs):
-        super(VADPerceptionTransformer, self).__init__(**kwargs)
-        self.encoder = build_transformer_layer_sequence(encoder)
-        if decoder is not None:
-            self.decoder = build_transformer_layer_sequence(decoder)
-        else:
-            self.decoder = None
-        if map_decoder is not None:
-            self.map_decoder = build_transformer_layer_sequence(map_decoder)
-        else:
-            self.map_decoder = None
-
-        self.embed_dims = embed_dims
-        self.num_feature_levels = num_feature_levels
-        self.num_cams = num_cams
-        self.fp16_enabled = False
-        self.rotate_prev_bev = rotate_prev_bev
-        self.use_shift = use_shift
-        self.use_can_bus = use_can_bus
-        self.can_bus_norm = can_bus_norm
-        self.use_cams_embeds = use_cams_embeds
-        self.two_stage_num_proposals = two_stage_num_proposals
-        self.rotate_center = rotate_center
-        self.map_num_vec = map_num_vec
-        self.map_num_pts_per_vec = map_num_pts_per_vec
-        self.init_layers()
-
-    def init_layers(self):
-        """Initialize layers of the Detr3DTransformer."""
-        self.level_embeds = nn.Parameter(torch.Tensor(
-            self.num_feature_levels, self.embed_dims))
-        self.cams_embeds = nn.Parameter(
-            torch.Tensor(self.num_cams, self.embed_dims))
-        self.reference_points = nn.Linear(self.embed_dims, 3)
-        self.map_reference_points = nn.Linear(self.embed_dims, 2)
-        self.can_bus_mlp = nn.Sequential(
-            nn.Linear(18, self.embed_dims // 2),
-            nn.ReLU(inplace=True),
-            nn.Linear(self.embed_dims // 2, self.embed_dims),
-            nn.ReLU(inplace=True),
-        )
-        if self.can_bus_norm:
-            self.can_bus_mlp.add_module('norm', nn.LayerNorm(self.embed_dims))
-
-    def init_weights(self):
-        """Initialize the transformer weights."""
-        for p in self.parameters():
-            if p.dim() > 1:
-                nn.init.xavier_uniform_(p)
-        for m in self.modules():
-            #if isinstance(m, MSDeformableAttention3D) or isinstance(m, TemporalSelfAttention) \
-            #        or isinstance(m, CustomMSDeformableAttention):
-            if isinstance(m, MultiScaleDeformableAttention3D) or isinstance(m, TemporalSelfAttention) \
-                    or isinstance(m, BEVMultiScaleDeformableAttention):
-                try:
-                    m.init_weight()
-                except AttributeError:
-                    m.init_weights()
-        normal_(self.level_embeds)
-        normal_(self.cams_embeds)
-        xavier_init(self.reference_points, distribution='uniform', bias=0.)
-        xavier_init(self.map_reference_points, distribution='uniform', bias=0.)
-        xavier_init(self.can_bus_mlp, distribution='uniform', bias=0.)
-
-    def get_bev_features(
-            self,
-            mlvl_feats,
-            bev_queries,
-            bev_h,
-            bev_w,
-            grid_length=[0.512, 0.512],
-            bev_pos=None,
-            prev_bev=None,
-            **kwargs):
-        """
-        obtain bev features.
-        """
-
-        bs = mlvl_feats[0].size(0)
-        bev_queries = bev_queries.unsqueeze(1).repeat(1, bs, 1)
-        bev_pos = bev_pos.flatten(2).permute(2, 0, 1)
-
-        # obtain rotation angle and shift with ego motion
-        delta_x = np.array([each['bev_attr'][0]
-                           for each in kwargs['img_metas']])
-        delta_y = np.array([each['bev_attr'][1]
-                           for each in kwargs['img_metas']])
-        ego_angle = np.array(
-            [each['bev_attr'][-2] / np.pi * 180 for each in kwargs['img_metas']])
-        grid_length_y = grid_length[0]
-        grid_length_x = grid_length[1]
-        translation_length = np.sqrt(delta_x ** 2 + delta_y ** 2)
-        translation_angle = np.arctan2(delta_y, delta_x) / np.pi * 180
-        bev_angle = ego_angle - translation_angle
-        shift_y = translation_length * \
-            np.cos(bev_angle / 180 * np.pi) / grid_length_y / bev_h
-        shift_x = translation_length * \
-            np.sin(bev_angle / 180 * np.pi) / grid_length_x / bev_w
-        shift_y = shift_y * self.use_shift
-        shift_x = shift_x * self.use_shift
-        shift = bev_queries.new_tensor(
-            [shift_x, shift_y]).permute(1, 0)  # xy, bs -> bs, xy
-
-        if prev_bev is not None:
-            if prev_bev.shape[1] == bev_h * bev_w:
-                prev_bev = prev_bev.permute(1, 0, 2)
-            if self.rotate_prev_bev:
-                for i in range(bs):
-                    # num_prev_bev = prev_bev.size(1)
-                    rotation_angle = kwargs['img_metas'][i]['bev_attr'][-1]
-                    tmp_prev_bev = prev_bev[:, i].reshape(
-                        bev_h, bev_w, -1).permute(2, 0, 1)
-                    tmp_prev_bev = rotate(tmp_prev_bev, float(rotation_angle),
-                                          center=self.rotate_center)
-                    tmp_prev_bev = tmp_prev_bev.permute(1, 2, 0).reshape(
-                        bev_h * bev_w, 1, -1)
-                    prev_bev[:, i] = tmp_prev_bev[:, 0]
-
-        # add can bus signals
-        can_bus = bev_queries.new_tensor(
-            [each['bev_attr'] for each in kwargs['img_metas']])  # [:, :]
-        can_bus = self.can_bus_mlp(can_bus)[None, :, :]
-        bev_queries = bev_queries + can_bus * self.use_can_bus
-
-        feat_flatten = []
-        spatial_shapes = []
-        for lvl, feat in enumerate(mlvl_feats):
-            bs, num_cam, c, h, w = feat.shape
-            spatial_shape = (h, w)
-            feat = feat.flatten(3).permute(1, 0, 3, 2)
-            if self.use_cams_embeds:
-                feat = feat + self.cams_embeds[:, None, None, :].to(feat.dtype)
-            feat = feat + self.level_embeds[None,
-                                            None, lvl:lvl + 1, :].to(feat.dtype)
-            spatial_shapes.append(spatial_shape)
-            feat_flatten.append(feat)
-
-        feat_flatten = torch.cat(feat_flatten, 2)
-        spatial_shapes = torch.as_tensor(
-            spatial_shapes, dtype=torch.long, device=bev_pos.device)
-        level_start_index = torch.cat((spatial_shapes.new_zeros(
-            (1,)), spatial_shapes.prod(1).cumsum(0)[:-1]))
-
-        feat_flatten = feat_flatten.permute(
-            0, 2, 1, 3)  # (num_cam, H*W, bs, embed_dims)
-
-        bev_embed = self.encoder(
-            bev_queries,
-            feat_flatten,
-            feat_flatten,
-            bev_h=bev_h,
-            bev_w=bev_w,
-            bev_pos=bev_pos,
-            spatial_shapes=spatial_shapes,
-            level_start_index=level_start_index,
-            prev_bev=prev_bev,
-            shift=shift,
-            **kwargs
-        )
-
-        return bev_embed
-
-    def forward(self,
-                mlvl_feats,
-                bev_queries,
-                object_query_embed,
-                map_query_embed,
-                bev_h,
-                bev_w,
-                grid_length=[0.512, 0.512],
-                bev_pos=None,
-                reg_branches=None,
-                cls_branches=None,
-                map_reg_branches=None,
-                map_cls_branches=None,                
-                prev_bev=None,            
-                **kwargs):
-        """Forward function for `Detr3DTransformer`.
-        Args:
-            mlvl_feats (list(Tensor)): Input queries from
-                different level. Each element has shape
-                [bs, num_cams, embed_dims, h, w].
-            bev_queries (Tensor): (bev_h*bev_w, c)
-            bev_pos (Tensor): (bs, embed_dims, bev_h, bev_w)
-            object_query_embed (Tensor): The query embedding for decoder,
-                with shape [num_query, c].
-            reg_branches (obj:`nn.ModuleList`): Regression heads for
-                feature maps from each decoder layer. Only would
-                be passed when `with_box_refine` is True. Default to None.
-        Returns:
-            tuple[Tensor]: results of decoder containing the following tensor.
-                - bev_embed: BEV features
-                - inter_states: Outputs from decoder. If
-                    return_intermediate_dec is True output has shape \
-                      (num_dec_layers, bs, num_query, embed_dims), else has \
-                      shape (1, bs, num_query, embed_dims).
-                - init_reference_out: The initial value of reference \
-                    points, has shape (bs, num_queries, 4).
-                - inter_references_out: The internal value of reference \
-                    points in decoder, has shape \
-                    (num_dec_layers, bs,num_query, embed_dims)
-                - enc_outputs_class: The classification score of \
-                    proposals generated from \
-                    encoder's feature maps, has shape \
-                    (batch, h*w, num_classes). \
-                    Only would be returned when `as_two_stage` is True, \
-                    otherwise None.
-                - enc_outputs_coord_unact: The regression results \
-                    generated from encoder's feature maps., has shape \
-                    (batch, h*w, 4). Only would \
-                    be returned when `as_two_stage` is True, \
-                    otherwise None.
-        """
-
-        bev_embed = self.get_bev_features(
-            mlvl_feats,
-            bev_queries,
-            bev_h,
-            bev_w,
-            grid_length=grid_length,
-            bev_pos=bev_pos,
-            prev_bev=prev_bev,
-            **kwargs)  # bev_embed shape: bs, bev_h*bev_w, embed_dims
-
-        bs = mlvl_feats[0].size(0)
-        query_pos, query = torch.split(
-            object_query_embed, self.embed_dims, dim=1)
-        query_pos = query_pos.unsqueeze(0).expand(bs, -1, -1)
-        query = query.unsqueeze(0).expand(bs, -1, -1)
-        reference_points = self.reference_points(query_pos)
-        reference_points = reference_points.sigmoid()
-        init_reference_out = reference_points
-
-        map_query_pos, map_query = torch.split(
-            map_query_embed, self.embed_dims, dim=1)
-        map_query_pos = map_query_pos.unsqueeze(0).expand(bs, -1, -1)
-        map_query = map_query.unsqueeze(0).expand(bs, -1, -1)
-        map_reference_points = self.map_reference_points(map_query_pos)
-        map_reference_points = map_reference_points.sigmoid()
-        map_init_reference_out = map_reference_points        
-
-        query = query.permute(1, 0, 2)
-        query_pos = query_pos.permute(1, 0, 2)
-        map_query = map_query.permute(1, 0, 2)
-        map_query_pos = map_query_pos.permute(1, 0, 2)
-        bev_embed = bev_embed.permute(1, 0, 2)
-
-        if self.decoder is not None:
-            # [L, Q, B, D], [L, B, Q, D]
-            inter_states, inter_references = self.decoder(
-                query=query,
-                key=None,
-                value=bev_embed,
-                query_pos=query_pos,
-                reference_points=reference_points,
-                reg_branches=reg_branches,
-                cls_branches=cls_branches,
-                spatial_shapes=torch.tensor([[bev_h, bev_w]], device=query.device),
-                level_start_index=torch.tensor([0], device=query.device),
-                **kwargs)
-            inter_references_out = inter_references
-        else:
-            inter_states = query.unsqueeze(0)
-            inter_references_out = reference_points.unsqueeze(0)
-
-        if self.map_decoder is not None:
-            # [L, Q, B, D], [L, B, Q, D]
-            map_inter_states, map_inter_references = self.map_decoder(
-                query=map_query,
-                key=None,
-                value=bev_embed,
-                query_pos=map_query_pos,
-                reference_points=map_reference_points,
-                reg_branches=map_reg_branches,
-                cls_branches=map_cls_branches,
-                spatial_shapes=torch.tensor([[bev_h, bev_w]], device=map_query.device),
-                level_start_index=torch.tensor([0], device=map_query.device),
-                **kwargs)
-            map_inter_references_out = map_inter_references
-        else:
-            map_inter_states = map_query.unsqueeze(0)
-            map_inter_references_out = map_reference_points.unsqueeze(0)
-
-        return (
-            bev_embed, inter_states, init_reference_out, inter_references_out,
-            map_inter_states, map_init_reference_out, map_inter_references_out)
-
-
-@MODELS.register_module()
-class CustomTransformerDecoder(TransformerLayerSequence):
-    """Implements the decoder in DETR3D transformer.
+class VADCustomDecoder(TransformerLayerSequence):
+    """Implements the decoders in VAD for agent-agent, agent-map, ego-agent, ego-map interactions.
     Args:
         return_intermediate (bool): Whether to return intermediate outputs.
         coder_norm_cfg (dict): Config of last normalization layer. Default: `LN`.
     """
 
     def __init__(self, *args, return_intermediate=False, **kwargs):
-        super(CustomTransformerDecoder, self).__init__(*args, **kwargs)
+        super(VADCustomDecoder, self).__init__(*args, **kwargs)
         self.return_intermediate = return_intermediate
-        self.fp16_enabled = False
 
     def forward(self,
                 query,
@@ -448,7 +129,7 @@ class CustomTransformerDecoder(TransformerLayerSequence):
                 key_padding_mask=None,
                 *args,
                 **kwargs):
-        """Forward function for `Detr3DTransformerDecoder`.
+        """Forward function
         Args:
             query (Tensor): Input query with shape
                 `(num_query, bs, embed_dims)`.
