+++ fsd/agents/VAD/__init__.py
@@ -1,7 +1,5 @@
 from .task_modules import CustomNMSFreeCoder, MapNMSFreeCoder
 
-from .VAD_transformer import (MapDetectionTransformerDecoder,
-                              VADPerceptionTransformer,
-                              CustomTransformerDecoder)
+from .VAD_transformer import (VectorMapDecoder, VADCustomDecoder)
 from .VAD_head import VADHead
 from .VAD import VAD
